Cambios en el hito3 y CLI
