+++ index.js
@@ -1,44 +1,97 @@
-#!/usr/bin/env node
-const mdLinks = require('./hito3-1.js'); // Reemplaza './mdLinks' con la ruta correcta a tu función mdLinks
-const argv = require('yargs').argv;
-const chalk = require('chalk'); // Importar la biblioteca chalk
+const fs = require('fs/promises');
+const path = require('path');
+const markdownIt = require('markdown-it');
+const encontrarLinks = require('./hito1');
+const validarLink = require('./hito2');
 
-const path = argv._[0];
-const validate = argv.validate || false;
-const stats = argv.stats || false;
+const mdLinks = (absolutaPath, validate) => {
+  return new Promise((resolve, reject) => {
+    // Convertir la ruta absoluta
+    const convertAbsoluta = path.resolve(absolutaPath);
 
-// Lógica para invocar mdLinks con los argumentos correctos
-mdLinks(path, validate)
-  .then((links) => {
-    if (stats) {
-      const totalLinks = links.length;
-      const uniqueLinks = [...new Set(links.map((link) => link.href))].length;
-      console.log(`Total de enlaces: ${totalLinks}`);
-      console.log(`Enlaces únicos: ${uniqueLinks}`);
-    } else {
-      links.forEach((link) => {
-        const statusMessage = validate ? (link.ok ? chalk.green('OK') : chalk.red('FAIL')) : '';
-        console.log(`Ruta: ${link.file}, Texto: ${link.text}, Enlace: ${link.href}, Estado: ${statusMessage}`);
+    // Verificamos si la ruta existe
+    fs.access(convertAbsoluta)
+      .then(() => {
+        // Leer la información del archivo o directorio
+        return fs.stat(convertAbsoluta);
+      })
+      .then((stats) => {
+        if (stats.isDirectory()) {
+          // Si es un directorio, leer los archivos dentro de él
+          return fs.readdir(convertAbsoluta);
+        } else if (stats.isFile() && path.extname(convertAbsoluta) === '.md') {
+          // Si es un archivo Markdown válido, leer su contenido
+          return fs.readFile(convertAbsoluta, 'utf-8')
+            .then((contenido) => {
+              const md = markdownIt();
+              const html = md.render(contenido);
+              const links = encontrarLinks(html, convertAbsoluta);
+
+              if (!validate) {
+                resolve(links);
+              } else {
+                const linkPromises = links.map((link) => {
+                  return validarLink(link);
+                });
+
+                Promise.all(linkPromises)
+                  .then((validados) => {
+                    resolve(validados);
+                  })
+                  .catch((error) => {
+                    reject(error);
+                  });
+              }
+            });
+        } else {
+          reject(new Error('El archivo/directorio no es válido.'));
+        }
+      })
+      .then((files) => {
+        // Si es un directorio, procesar los archivos .md dentro de él
+        if (Array.isArray(files)) {
+          const linkPromises = [];
+
+          files.forEach((file) => {
+            const filePath = path.join(convertAbsoluta, file);
+            const extname = path.extname(filePath);
+
+            if (extname === '.md') {
+              linkPromises.push(
+                fs.readFile(filePath, 'utf-8')
+                  .then((contenido) => {
+                    const md = markdownIt();
+                    const html = md.render(contenido);
+                    const links = encontrarLinks(html, filePath);
+
+                    if (!validate) {
+                      return links;
+                    } else {
+                      const validatedLinks = links.map((link) => {
+                        return validarLink(link);
+                      });
+
+                      return Promise.all(validatedLinks);
+                    }
+                  })
+              );
+            }
+          });
+
+          Promise.all(linkPromises)
+            .then((results) => {
+              const allLinks = validate ? [].concat(...results) : results;
+              resolve(allLinks);
+            })
+            .catch((error) => {
+              reject(error);
+            });
+        }
+      })
+      .catch((error) => {
+        reject(error);
       });
-    }
-  })
-  .catch((error) => {
-    // Colorear mensajes de error en rojo
-    console.error(chalk.red('Error:', error.message));
   });
+};
 
-
-
-/*const  mdLinks = require('./hito3-1.js');
-
-//mdLinks('./fileMD/archivo.txt')
-mdLinks('/Users/LENOVO/DEV009-md-links/',true)
-.then(links => {
-  console.log(links);
-})
-.catch(error => {
-  console.error(error.message);
-});*/
-
-// llamamos la funcion de md-links
-
+module.exports = mdLinks;
