+++ package.json
@@ -2,6 +2,7 @@
   "name": "abril-mdlinks",
   "version": "0.1.0",
   "description": "md-links CLI",
+  "main": "index.js",
   "author": "Abril Vargas",
   "keywords": [
     "javascript",
@@ -32,6 +33,7 @@
     "axios-mock-adapter": "^1.21.5",
     "chalk": "^4.1.2",
     "cheerio": "^1.0.0-rc.12",
+    "cli-boxes": "^3.0.0",
     "figlet": "^1.6.0",
     "markdown-it": "^13.0.1",
     "yargs": "^17.7.2"
